Well, I never actually implemented the unit radioButton of miles/kilos... This is the same issue I came accross in Sprint1 (I commented out this style of code b ack then too, because it cannot be done smoothly, and google doesn't provide help). If anything, we can revert back to a dropdown menu (way easier to manage).
The issue I am having deals with the "spread" operator="..." and nested state objects are immutable. Not sure if I can change directly, or have to re-do the trip TFFI with a deep-copy and this.setState({trip: myDeepCopyObj}).

Anyways, the actual UI is not visually functional, nor does it officially change state.