1.) Created a REST API for Query, called by a HTTP post: /query (look at how /plan is posted if youre confused The flow is somewhat like this: Microserver gets a post that is either plan or query, based on this it will either call the method plan() or query(). Plan(), as we know, makes a plan object with the request (json string) then converts it to Java object Trip using the .getTrip() method. Query() does something similar. It creates a search object with the request, then converts to a Java object Query using .getQuery() method. The naming is not exactly analogous between the two, but if you get past that it makes sense.
2.) Created the query and search java classes
* Search.java
As noted in the java doc:  This class handles to the conversions of the query request/response, converting from the Json string in the request body to a Query object, searching the DB using the Query, and converting the resulting Query object back to a Json string so it may returned as the response.

* Query.java
This class *does* nothing, but it does have a method searchDB which will do some magic and talk to the DB once its implemented.
It also has fields for all of the query fields potentially present in the query JSON:
String type, String query, ArrayList<Place> places.